refactor (Tree): reduce complexity
