+++ src/components/OrderableList/types.ts
@@ -2,7 +2,7 @@
 
 import { ReactElement } from 'react';
 import { DraggableItem } from '@utilities/dnd';
-import { TreeItemStyling } from '..';
+import { DragHandlerPosition, TreeItemStyling } from '..';
 
 export type RenderListItem<T> = (items: OrderableListItem<T>) => ReactElement;
 
@@ -15,7 +15,7 @@ export type OrderableListItemStyle = TreeItemStyling;
 export type OrderableListProps<T> = {
     items: OrderableListItem<T>[];
     dragDisabled: boolean;
-    dragHandlerPosition?: 'left' | 'right' | 'none';
+    dragHandlerPosition?: DragHandlerPosition;
     itemStyle?: OrderableListItemStyle;
     selectedId?: string;
     onMove: (modifiedItems: DraggableItem<T>[]) => void;
